v0.3.1025 - /topics leaves client.ts
SCALE-SEAM ⑳. Seven BCF methods (create, viewpoints, board, timeline,
comments) move to api/topics.ts. Three regions. pins() stays on /pins.
client.ts 3,243 → 3,205.
