+++ services/api/test_file_sizes.py
@@ -72,7 +72,7 @@ PER_FILE = {
     #: other modules and one got a field in only by appending to an existing line. That is the
     #: friction this ratchet is for, and it asked the intended question — the contract documents had
     #: a whole domain to leave with, so the number goes DOWN by 32 instead of up by one.
-    "apps/web/src/api/client.ts": 3_243,   # 22 SCALE-SEAM ⑲ /mep -> mep.ts (3,304 -> 3,243)   # 21 SCALE-SEAM ⑱ /documents -> documents.ts (3,353 -> 3,304)   # ⑳ SCALE-SEAM ⑰ /models -> models.ts (3,412 -> 3,353)   # ⑲ SCALE-SEAM ⑯ /elements -> elements.ts (3,482 -> 3,412)   # ⑱ SCALE-SEAM ⑮ /drawings -> drawingSheets.ts (3,538 -> 3,482)   # ⑰ SCALE-SEAM ⑭ /drawing-set -> drawingSet.ts (3,602 -> 3,538)   # ⑯ SCALE-SEAM ⑬ /sync -> sync.ts (3,629 -> 3,602)   # ⑮ SCALE-SEAM ⑫ /connections -> connections.ts (3,672 -> 3,629)   # ⑭ camStatementUrl left — CAM PDF is POST from api/downloadPdf.ts (3,679 -> 3,672)   # + cost group -> cost.ts;   # +1 = the withRoutines IMPORT, not an endpoint:   # ⑨ contracts -> contracts.ts; ⑧ 3,796; ⑦ 3,871; before ⑦ 3,967   # + ⑩ finance -> finance.ts (both extractions are in the merged file)   # ⑪ doc-QA -> docqa.ts: adding ONE method for R31-CITE-HIGHLIGHT hit 3,715 and failed here, which is this pin working; the review* cluster + its helper moved out instead, so 3,703 -> 3,690; ⑫ the BidLevelingDetail RESPONSE TYPE -> types.ts: widening it for `non_responsive` hit 3,694 and failed here, which is the pin working -- a DTO belongs in types.ts, so 3,690 -> 3,685   # ⑬ the SAME move again: adding `entitlementReviewCycles` for R22-ENTITLEMENT hit 3,706 and failed here. Both that response type AND the `diligenceReadiness` one beside it went to types.ts, so 3,685 -> 3,683 -- the pin bought a SECOND DTO out of the file rather than merely admitting the first
+    "apps/web/src/api/client.ts": 3_205,   # 23 SCALE-SEAM ⑳ /topics -> topics.ts (3,243 -> 3,205)   # 22 SCALE-SEAM ⑲ /mep -> mep.ts (3,304 -> 3,243)   # 21 SCALE-SEAM ⑱ /documents -> documents.ts (3,353 -> 3,304)   # ⑳ SCALE-SEAM ⑰ /models -> models.ts (3,412 -> 3,353)   # ⑲ SCALE-SEAM ⑯ /elements -> elements.ts (3,482 -> 3,412)   # ⑱ SCALE-SEAM ⑮ /drawings -> drawingSheets.ts (3,538 -> 3,482)   # ⑰ SCALE-SEAM ⑭ /drawing-set -> drawingSet.ts (3,602 -> 3,538)   # ⑯ SCALE-SEAM ⑬ /sync -> sync.ts (3,629 -> 3,602)   # ⑮ SCALE-SEAM ⑫ /connections -> connections.ts (3,672 -> 3,629)   # ⑭ camStatementUrl left — CAM PDF is POST from api/downloadPdf.ts (3,679 -> 3,672)   # + cost group -> cost.ts;   # +1 = the withRoutines IMPORT, not an endpoint:   # ⑨ contracts -> contracts.ts; ⑧ 3,796; ⑦ 3,871; before ⑦ 3,967   # + ⑩ finance -> finance.ts (both extractions are in the merged file)   # ⑪ doc-QA -> docqa.ts: adding ONE method for R31-CITE-HIGHLIGHT hit 3,715 and failed here, which is this pin working; the review* cluster + its helper moved out instead, so 3,703 -> 3,690; ⑫ the BidLevelingDetail RESPONSE TYPE -> types.ts: widening it for `non_responsive` hit 3,694 and failed here, which is the pin working -- a DTO belongs in types.ts, so 3,690 -> 3,685   # ⑬ the SAME move again: adding `entitlementReviewCycles` for R22-ENTITLEMENT hit 3,706 and failed here. Both that response type AND the `diligenceReadiness` one beside it went to types.ts, so 3,685 -> 3,683 -- the pin bought a SECOND DTO out of the file rather than merely admitting the first
     #: R39-DECOMP-VIEWER. Pinned at its CURRENT size before any extraction, deliberately.
     #:
     #: `app.ts` had no per-file entry, so it lived under the 5,200 global — which it also *set*, being
